feat(US/24): config de envio de email
